+++ tests/utils.py
@@ -1,10 +1,8 @@
 import numpy as np
 from scipy.spatial import Delaunay
 
-from coxeter.shapes import ConvexPolyhedron
 
-
-def compute_inertia_mc(vertices, num_samples=1e6):
+def compute_inertia_mc(vertices, volume, num_samples=1e6):
     """Compute inertia tensor via Monte Carlo integration.
 
     Using Monte Carlo integration provides a means to test the results of an
@@ -14,7 +12,7 @@ def compute_inertia_mc(vertices, num_samples=1e6):
         num_samples (int): The number of samples to use.
 
     Returns:
-        float: The 3x3 inertia tensor.
+        (3, 3) :class:`numpy.ndarray`: The 3x3 inertia tensor.
     """
     mins = np.min(vertices, axis=0)
     maxs = np.max(vertices, axis=0)
@@ -31,11 +29,37 @@ def compute_inertia_mc(vertices, num_samples=1e6):
     i_xz = np.mean(-points[inside][:, 0] * points[inside][:, 2])
     i_yz = np.mean(-points[inside][:, 1] * points[inside][:, 2])
 
-    poly = ConvexPolyhedron(vertices)
-
     inertia_tensor = (
-        np.array([[i_xx, i_xy, i_xz], [i_xy, i_yy, i_yz], [i_xz, i_yz, i_zz]])
-        * poly.volume
+        np.array([[i_xx, i_xy, i_xz], [i_xy, i_yy, i_yz], [i_xz, i_yz, i_zz]]) * volume
     )
+
+    return inertia_tensor
+
+
+def compute_centroid_mc(vertices, num_samples=1e6):
+    """Compute centroid via Monte Carlo integration.
+
+    Using Monte Carlo integration provides a means to test the results of an
+    analytical calculation.
+
+    Args:
+        num_samples (int): The number of samples to use.
+
+    Returns:
+        (3, ) :class:`numpy.ndarray`: The center.
+    """
+    mins = np.min(vertices, axis=0)
+    maxs = np.max(vertices, axis=0)
+
+    points = np.random.rand(int(num_samples), 3) * (maxs - mins) + mins
+
+    hull = Delaunay(vertices)
+    inside = hull.find_simplex(points) >= 0
+
+    c_x = np.mean(points[inside][:, 0])
+    c_y = np.mean(points[inside][:, 1])
+    c_z = np.mean(points[inside][:, 2])
+
+    inertia_tensor = np.array([c_x, c_y, c_z])
 
     return inertia_tensor